% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brca_split_dataset.R
\name{brca_split_dataset}
\alias{brca_split_dataset}
\title{Function of dividing BRCA data set}
\usage{
brca_split_dataset(data, seed = 900)
}
\arguments{
\item{data}{: dataset}

\item{seed}{:random seed}
}
\value{
:Returns a list of training sets, test sets and verification sets.
}
\description{
Function of dividing BRCA data set
}
\examples{
DHMOC.brca_split_dataset(data,seed)
}
