% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_split_dataset.R
\name{gc_split_dataset}
\alias{gc_split_dataset}
\title{Function of dividing GC data set}
\usage{
gc_split_dataset(data, seed = 123)
}
\arguments{
\item{data}{: dataset}

\item{seed}{:random seed}
}
\value{
:Returns a list of training sets, test sets and verification sets.
}
\description{
Function of dividing GC data set
}
\examples{
DHMOC.gc_split_dataset(data,seed)
}
