% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCC_split_dataset.R
\name{HCC_split_dataset}
\alias{HCC_split_dataset}
\title{Function of dividing spatial transcriptome data of hepatocellular carcinoma (HCC)}
\usage{
HCC_split_dataset(data, seed = 123)
}
\arguments{
\item{data}{:Spatial Transcriptome Data of Hepatocellular Carcinoma (HCC)}

\item{seed}{:random seed}
}
\value{
:Contains a list of training sets, test sets and verification sets.
}
\description{
Function of dividing spatial transcriptome data of hepatocellular carcinoma (HCC)
}
\examples{
DHMOC.HCC_split_dataset(data,seed)
}
