% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktuple_split_dataset.R
\name{ktuple_split_dataset}
\alias{ktuple_split_dataset}
\title{Function of dividing Faecal data set}
\usage{
ktuple_split_dataset(data, seed = 123)
}
\arguments{
\item{data}{: dataset}

\item{seed}{:random seed}
}
\value{
:Returns a list of training sets, test sets and verification sets.
}
\description{
Function of dividing Faecal data set
}
\examples{
DHMOC.brca_split_dataset(data,seed)
}
