% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm_run_ovr_models.R
\name{lightgbm_run_ovr_models}
\alias{lightgbm_run_ovr_models}
\title{Lightgbm model}
\usage{
lightgbm_run_ovr_models(train_data, test_data)
}
\arguments{
\item{train_data}{:Training data set}

\item{test_data}{:Testing data set}
}
\value{
:List containing prediction labels and confusion matrices
}
\description{
Lightgbm model
}
\examples{
DHMOC.lightgbm_run_ovr_models(train,test)
}
