% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm_train_test_ovr_model.R
\name{lightgbm_train_test_ovr_model}
\alias{lightgbm_train_test_ovr_model}
\title{Lightgbm training test model}
\usage{
lightgbm_train_test_ovr_model(train_data, test_data, positive_class)
}
\arguments{
\item{train_data}{:training dataset}

\item{test_data}{:testing dataset}

\item{positive_class}{:Positive label}
}
\value{
:Returns the prediction label of the test.
}
\description{
Lightgbm training test model
}
\examples{
DHMOC.lightgbm_train_test_ovr_model(train,test,class)
}
