% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lymphoid_split_dataset.R
\name{lymphoid_split_dataset}
\alias{lymphoid_split_dataset}
\title{Function of single cell RNA sequence data set for dividing lymphocytes}
\usage{
lymphoid_split_dataset(data, seed = 123)
}
\arguments{
\item{data}{:Single cell RNA sequence data set of lymphocytes}

\item{seed}{:random seed}
}
\value{
:Contains a list of verification sets, training sets and test sets.
}
\description{
Function of single cell RNA sequence data set for dividing lymphocytes
}
\examples{
DHMOC.lymphoid_split_dataset(data,123)
}
