% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCIrna_split_dataset.R
\name{NCIrna_split_dataset}
\alias{NCIrna_split_dataset}
\title{Function of dividing single cell RNA sequence of gastric cancer cell line NCI-N87}
\usage{
NCIrna_split_dataset(data, seed = 123)
}
\arguments{
\item{data}{:Single cell RNA sequence of gastric cancer cell line NCI-N87}

\item{seed}{:random seed}
}
\value{
:Returns a data list containing training set, test set and verification set.
}
\description{
Function of dividing single cell RNA sequence of gastric cancer cell line NCI-N87
}
\examples{
DHMOC.NCIrna_split_dataset(data,seed)
}
