% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ovr_models.R
\name{run_ovr_models}
\alias{run_ovr_models}
\title{Function to run the model}
\usage{
run_ovr_models(train_data, test_data)
}
\arguments{
\item{train_data}{:Training data set}

\item{test_data}{: Testing data set}
}
\value{
: Returns a list containing training labels, test labels and confusion matrices.
}
\description{
Function to run the model
}
\examples{
DHMOC.run_ovr_models(train,test)
}
