% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_model.R
\name{svm_model}
\alias{svm_model}
\title{SVM model}
\usage{
svm_model(train_data, test_data)
}
\arguments{
\item{train_data}{:Training dataset}

\item{test_data}{:Testing data set}
}
\value{
:Returns a list of training sets, test sets, test labels and mixing matrices.
}
\description{
SVM model
}
\examples{
DHMOC.svm_model(train,test)
}
