% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_ovr_model.R
\name{train_test_ovr_model}
\alias{train_test_ovr_model}
\title{Training set and test set model}
\usage{
train_test_ovr_model(train_data, test_data, positive_class)
}
\arguments{
\item{train_data}{:training dataset}

\item{test_data}{:testing dataset}

\item{positive_class}{:}
}
\value{
:Returns a list containing training labels and test labels.
}
\description{
Training set and test set model
}
\examples{
DHMOC.train_test_ovr_model(train,test,calss)
}
